+++ packages/components/src/public-api.ts
@@ -17,6 +17,7 @@ export * from './color-picker/index';
 export * from './combobox/index';
 export * from './confirm-popup/index';
 export * from './date-picker/index';
+export * from './datetime-picker/index';
 export * from './dialog/index';
 export * from './drawer/index';
 export * from './dropdown-menu/index';